# docs/weixin-upstream-openapi-3.1.yaml
openapi: 3.1.0
info:
  title: Weixin Upstream Communication API
  version: 0.1.0
  summary: weixin-local-chat 依赖的微信上游接口
  description: >
    根据 src/api/weixin.ts 的实际调用逻辑整理。实际基础地址由本地配置 baseUrl
    提供，默认值为 https://ilinkai.weixin.qq.com。
servers:
  - url: https://ilinkai.weixin.qq.com
    description: 默认微信上游地址
...
paths:
...
  /ilink/bot/getconfig:
    post:
      tags:
        - Config
      summary: 获取会话配置
      operationId: getConfig
      parameters:
        - $ref: "#/components/parameters/AuthorizationTypeHeader"
        - $ref: "#/components/parameters/AuthorizationHeader"
        - $ref: "#/components/parameters/WechatUinHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetConfigRequest"
      responses:
        "200":
          description: 会话配置
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetConfigResponse"
        "500":
          $ref: "#/components/responses/UpstreamError"
...
components:
  parameters:
...
    AuthorizationTypeHeader:
      name: AuthorizationType
      in: header
      required: true
      description: 当前固定值为 ilink_bot_token
      schema:
        type: string
        example: ilink_bot_token
    AuthorizationHeader:
      name: Authorization
      in: header
      required: true
      description: Bearer 机器人 token
      schema:
        type: string
        example: Bearer bot-token
    WechatUinHeader:
      name: X-WECHAT-UIN
      in: header
      required: true
      description: 随机生成的 base64 字符串
      schema:
        type: string
        example: MTIzNDU2Nzg=
  responses:
    UpstreamError:
      description: 上游接口异常
      content:
        text/plain:
          schema:
            type: string
  schemas:
    BaseInfo:
      type: object
      properties:
        channel_version:
          type: string
          example: weixin-local-chat/0.1.0
      required:
        - channel_version
...
    GetConfigRequest:
      type: object
      properties:
        ilink_user_id:
          type: string
        context_token:
          type: string
        base_info:
          $ref: "#/components/schemas/BaseInfo"
      required:
        - ilink_user_id
        - context_token
        - base_info
    GetConfigResponse:
      type: object
      properties:
        typing_ticket:
          type: string
      additionalProperties: true
...